requirements.txt
numpy
pytesseract
opencv-contrib-python
pillow